abstracted out the user dropdown component
