+++ client/src/components/layout/header/index.css
@@ -1,41 +1,3 @@
-.profile-button {
-    border-radius: 50%;
-    border-color: transparent;
-    background-color: transparent;
-}
-
-.profile-button:hover,
-.profile-button:focus,
-.profile-button:active {
-    border-radius: 50%;
-    border-color: transparent;
-    background-color: var(--secondary-color-hover);
-}
-
-.profile-icon,
-.profile-icon:focus {
-    height: 40px;
-    width: 40px;
-    color: var(--primary-color);
-    background-color: transparent;
-    font-weight: lighter;
-}
-
-.user-dropdown-link:link:focus, 
-.user-dropdown-link:link:hover,
-.user-dropdown-link:link:active {
-    margin-left: 10%;
-    background-color: #dadada;
-}
-
 .site-logo {
     cursor: pointer;
-}
-
-.user-dropdown-item,
-.user-dropdown:focus,
-.user-dropdown:active,
-.user-dropdown-item:hover {
-    color: black;
-    background-color: transparent;
 }
