+++ client/src/components/userDropdown/index.css
@@ -1,0 +1,30 @@
+.profile-button {
+    border-radius: 50%;
+    border-color: transparent;
+    background-color: transparent;
+}
+
+.profile-button:hover,
+.profile-button:focus,
+.profile-button:active {
+    border-radius: 50%;
+    border-color: transparent;
+    background-color: var(--secondary-color-hover);
+}
+
+.profile-icon,
+.profile-icon:focus {
+    height: 40px;
+    width: 40px;
+    color: var(--primary-color);
+    background-color: transparent;
+    font-weight: lighter;
+}
+
+.user-dropdown-item,
+.user-dropdown-item:focus,
+.user-dropdown-item:active,
+.user-dropdown-item:hover {
+    color: black;
+    background-color: transparent;
+}
